requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
faiss-cpu==1.11.0.post1
filelock==3.18.0
fsspec==2025.7.0
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
Pygments==2.19.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
typing_extensions==4.14.1
urllib3==2.5.0
